[refactor] split up partition_mapping.py
