+++ python_modules/dagster-graphql/dagster_graphql/schema/partition_mappings.py
@@ -1,5 +1,5 @@
 import graphene
-from dagster._core.definitions.partitions.mapping.partition_mapping import PartitionMapping
+from dagster._core.definitions.partitions.mapping.base import PartitionMapping
 
 
 class GraphenePartitionMapping(graphene.ObjectType):
