+++ python_modules/dagster-test/dagster_test/toys/partitioned_assets/different_static_partitions.py
@@ -1,5 +1,6 @@
-from dagster import AssetIn, StaticPartitionMapping, asset
+from dagster import AssetIn, asset
 from dagster._core.definitions.partitions.definition.static import StaticPartitionsDefinition
+from dagster._core.definitions.partitions.mapping.static import StaticPartitionMapping
 
 
 @asset(partitions_def=StaticPartitionsDefinition(["a", "b", "c"]))
