+++ python_modules/dagster/dagster/__init__.py
@@ -363,15 +363,29 @@ from dagster._core.definitions.partitions.definition.time_window_subclasses impo
     MonthlyPartitionsDefinition as MonthlyPartitionsDefinition,
     WeeklyPartitionsDefinition as WeeklyPartitionsDefinition,
 )
-from dagster._core.definitions.partitions.mapping.partition_mapping import (
+from dagster._core.definitions.partitions.mapping.all import (
     AllPartitionMapping as AllPartitionMapping,
-    DimensionPartitionMapping as DimensionPartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.base import PartitionMapping as PartitionMapping
+from dagster._core.definitions.partitions.mapping.identity import (
     IdentityPartitionMapping as IdentityPartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.last import (
     LastPartitionMapping as LastPartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.multi.base import (
+    DimensionPartitionMapping as DimensionPartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.multi.multi_to_multi import (
     MultiPartitionMapping as MultiPartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.multi.multi_to_single import (
     MultiToSingleDimensionPartitionMapping as MultiToSingleDimensionPartitionMapping,
-    PartitionMapping as PartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.specific_partitions import (
     SpecificPartitionsPartitionMapping as SpecificPartitionsPartitionMapping,
+)
+from dagster._core.definitions.partitions.mapping.static import (
     StaticPartitionMapping as StaticPartitionMapping,
 )
 from dagster._core.definitions.partitions.mapping.time_window_partition_mapping import (
