+++ python_modules/dagster/dagster/_core/definitions/asset_dep.py
@@ -5,10 +5,8 @@ import dagster._check as check
 from dagster._annotations import PublicAttr
 from dagster._core.definitions.asset_check_spec import AssetCheckKey
 from dagster._core.definitions.events import AssetKey, CoercibleToAssetKey
-from dagster._core.definitions.partitions.mapping.partition_mapping import (
-    PartitionMapping,
-    warn_if_partition_mapping_not_builtin,
-)
+from dagster._core.definitions.partitions.mapping.base import PartitionMapping
+from dagster._core.definitions.partitions.utils.mapping import warn_if_partition_mapping_not_builtin
 from dagster._core.errors import DagsterInvalidDefinitionError, DagsterInvariantViolationError
 from dagster._utils.warnings import deprecation_warning
