+++ python_modules/dagster/dagster/_core/definitions/asset_in.py
@@ -10,7 +10,7 @@ from dagster._core.definitions.events import (
 )
 from dagster._core.definitions.input import NoValueSentinel
 from dagster._core.definitions.metadata import ArbitraryMetadataMapping
-from dagster._core.definitions.partitions.mapping.partition_mapping import PartitionMapping
+from dagster._core.definitions.partitions.mapping.base import PartitionMapping
 from dagster._core.types.dagster_type import DagsterType, resolve_dagster_type
