+++ python_modules/dagster/dagster/_core/definitions/data_version.py
@@ -713,11 +713,11 @@ class CachingStaleStatusResolver:
     def _get_partition_dependencies(
         self, *, key: "AssetKeyPartitionKey"
     ) -> Sequence["AssetKeyPartitionKey"]:
-        from dagster import AllPartitionMapping
         from dagster._core.definitions.events import AssetKeyPartitionKey
         from dagster._core.definitions.partitions.definition.time_window import (
             TimeWindowPartitionsDefinition,
         )
+        from dagster._core.definitions.partitions.mapping.all import AllPartitionMapping
 
         asset_deps = self.asset_graph.get(key.asset_key).parent_keys
