+++ python_modules/dagster/dagster_tests/asset_defs_tests/partition_mapping_tests/test_static_partition_mapping.py
@@ -1,6 +1,6 @@
 import pytest
-from dagster import StaticPartitionMapping
 from dagster._core.definitions.partitions.definition.static import StaticPartitionsDefinition
+from dagster._core.definitions.partitions.mapping.static import StaticPartitionMapping
 from dagster._core.definitions.partitions.subset.default import DefaultPartitionsSubset
 from dagster_shared.serdes import deserialize_value, serialize_value
