+++ python_modules/dagster/dagster_tests/declarative_automation_tests/legacy_tests/updated_scenarios/partition_scenarios.py
@@ -1,17 +1,12 @@
 import datetime
 
-from dagster import (
-    AssetDep,
-    AutoMaterializePolicy,
-    AutoMaterializeRule,
-    DimensionPartitionMapping,
-    IdentityPartitionMapping,
-    MultiPartitionMapping,
-    TimeWindowPartitionMapping,
-)
+from dagster import AssetDep, AutoMaterializePolicy, AutoMaterializeRule, TimeWindowPartitionMapping
 from dagster._core.definitions.auto_materialize_rule_impls import (
     DiscardOnMaxMaterializationsExceededRule,
 )
+from dagster._core.definitions.partitions.mapping.identity import IdentityPartitionMapping
+from dagster._core.definitions.partitions.mapping.multi.base import DimensionPartitionMapping
+from dagster._core.definitions.partitions.mapping.multi.multi_to_multi import MultiPartitionMapping
 from dagster._core.definitions.timestamp import TimestampWithTimezone
 from dagster._record import copy
