+++ python_modules/dagster/dagster_tests/storage_tests/branching_io_manager_tests/test_partitioned_branching_io_manager.py
@@ -9,12 +9,12 @@ from dagster import (
     DagsterInstance,
     Definitions,
     In,
-    StaticPartitionMapping,
     asset,
     job,
     op,
 )
 from dagster._core.definitions.partitions.definition.static import StaticPartitionsDefinition
+from dagster._core.definitions.partitions.mapping.static import StaticPartitionMapping
 from dagster._core.execution.context.compute import OpExecutionContext
 from dagster._core.storage.branching.branching_io_manager import BranchingIOManager
